Extract TimerConfigPanelRow into its own file
Move TimerConfigPanelRow, TimerRow, TIMER_ROWS, and formatDuration
into a dedicated file per the one-component-per-file convention.
